// standard_reports/item-list/latest/convert_data_js/src/generated-types/arguments.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type Arguments = ItemListFilters;

export interface ItemListFilters {
  /**
   * Item Code
   */
  itemCode?: string | null;
  /**
   * Item Name
   */
  itemName?: string | null;
  /**
   * Master List
   */
  masterListId?: string | null;
  /**
   * Is Active
   */
  isActive?: boolean | null;
  /**
   * Only out of stock
   */
  onlyOutOfStock?: boolean;
  /**
   * VEN Category
   */
  venCategory?: null | "V" | "E" | "N" | "V_OR_E" | "V_E_N" | "NONE";
  /**
   * sort by
   */
  sort?: "name" | "code" | "SOH" | "locationName" | "categoryName";
  /**
   * sort by dir
   */
  dir?: null | "asc" | "desc";
  [k: string]: unknown;
}